2019|10|08

+++ 03_Django/04_django_crud_review/articles/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, redirect
 from django.core.exceptions import ValidationError
-from .models import Article
+from .models import Article, Comment
+from IPython import embed
 
 def index(request):
     #articles = Article.objects.all()[::-1]
@@ -10,20 +11,16 @@ def index(request):
     context = {'articles': articles}
     return render(request, 'articles/index.html', context)
 
-def new(request):
-    return render(request, 'articles/new.html')
-
 def create(request):
-    try:
+    # POST 요청일 떄
+    if request.method == 'POST':
         title = request.POST.get('title')
         content = request.POST.get('content')
         article = Article(title=title, content=content)
-        article.full_clean()
-    except ValidationError:
-        raise ValidationError('Your Error Message')
-    else:
         article.save()
         return redirect('articles:detail', article.pk)
+    # GET 요청일 때
+    return render(request, 'articles/create.html')
     """
     title = request.POST.get('title')
     content = request.POST.get('content')
@@ -44,24 +41,40 @@ def create(request):
     return redirect(f'/articles/{article.pk}')
     """
 
-def detail(request, pk):
-    article = Article.objects.get(pk=pk)
-    context = {'article': article}
+def detail(request, article_pk):
+    article = Article.objects.get(pk=article_pk)
+    comments = article.comments.all()
+    context = {'article': article, 'comments': comments, }
     return render(request, 'articles/detail.html', context)
 
-def delete(request, pk):
-    article = Article.objects.get(pk=pk)
-    article.delete()
-    return redirect('articles:index')
+def delete(request, article_pk):
+    article = Article.objects.get(pk=article_pk)
+    if request.method == 'POST':
+        article.delete()
+        return redirect('articles:index')
+    return redirect('articles:detail', article.pk)
 
-def edit(request, pk):
-    article = Article.objects.get(pk=pk)
+def update(request, article_pk):
+    # embed()
+    article = Article.objects.get(pk=article_pk)
+    if request.method == 'POST':
+        article.title = request.POST.get('title')
+        article.content = request.POST.get('content')
+        article.save()
+        return redirect('articles:detail', article.pk)
     context = {'article': article}
-    return render(request, 'articles/edit.html', context)
+    return render(request, 'articles/update.html', context)
 
-def update(request, pk):
-    article = Article.objects.get(pk=pk)
-    article.title = request.POST.get('title')
-    article.content = request.POST.get('content')
-    article.save()
+def comments_create(request, article_pk):
+    article = Article.objects.get(pk=article_pk)
+    if request.method == 'POST':
+        comment_content = request.POST.get('comment_content')
+        comment = Comment(content=comment_content, article_id=article.pk)
+        comment.save()
+    return redirect('articles:detail', article.pk)
+
+def comments_delete(request, comment_pk):
+    comment = Comment.objects.get(pk=comment_pk)
+    article = comment.article
+    comment.delete()
     return redirect('articles:detail', article.pk)
